+++ src/b2aiprep/prepare/constants.py
@@ -121,9 +121,35 @@ SPEECH_TASKS = (
     "Word-color Stroop",
 )
 
-ALLOWED_COLUMNS: t.List[str] = json.loads(
-    files("b2aiprep").joinpath("prepare").joinpath("resources", "allowed_columns.json").read_text()
-)
+def _load_participant_exclusions() -> t.List[str]:
+    """Load the participant IDs to exclude from the dataset.
+
+    Returns
+    -------
+    list
+        The participant IDs to exclude.
+    """
+    b2ai_resources = files("b2aiprep").joinpath("prepare").joinpath("resources")
+    if b2ai_resources.joinpath("participant_id_to_exclude_v2.json").exists():
+        return json.loads(b2ai_resources.joinpath("participant_id_to_exclude_v2.json").read_text())
+    return json.loads(b2ai_resources.joinpath("participant_id_to_exclude.json").read_text())
+
+PARTICIPANT_ID_TO_REMOVE: t.List[str] = _load_participant_exclusions()
+
+def _load_audio_filestem_exclusions() -> t.List[str]:
+    """Load the audio filestems to exclude from the dataset.
+
+    Returns
+    -------
+    list
+        The audio filestems to exclude.
+    """
+    b2ai_resources = files("b2aiprep").joinpath("prepare").joinpath("resources")
+    if b2ai_resources.joinpath("audio_filestem_to_exclude.json").exists():
+        return json.loads(b2ai_resources.joinpath("audio_filestem_to_exclude.json").read_text())
+    return []
+
+AUDIO_FILESTEMS_TO_REMOVE: t.List[str] = _load_audio_filestem_exclusions()
 
 class Instrument(BaseModel):
     """Instruments are associated with fixed sets of columns and a string